[ADD] Dependant modules for chart of taxes port.
NOTE: This commit is simply to keep the ball rolling. Eventually we
will need to add these 3 modules to a manifest of some sort which shows them
as dependencies. We will also need to have them bundled in automatically when
downloading the account_mtd module as a zip from the app store. We also need to
build a list of which dependent modules have which licenses - which can then
aid us in determining which license the account_mtd module should have.